// src/modules/client/posts/models/banner.model.ts
export interface Banner {
    _id: string;
    imagePath: string;
    type: string;
    isActive: boolean;
    order: number;
    link?: string;
    title?: string;
    description?: string;
    position?: string;
    createdAt: string;
    updatedAt: string;
  }
  
  export interface BannerResponse {
    data: Banner[];
    total: number;
    page: number;
    totalPages: number;
  }